pyproject.toml
[tool.poetry]
name = "simbir-health"
version = "0.1.0"
description = ""
authors = ["Егор <despadino@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.31.0"
fastapi = "^0.115.0"
loguru = "^0.7.2"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
yarl = "^1.14.0"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
jwt = "^1.3.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2-binary = "^2.9.10"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
